Data.IndexedListLiterals

Documentation

class IndexedListLiterals (input :: Type) (length :: Nat) (output :: Type) | output length -> input, input -> output length where

A type class which allows you to write tuples which can be transformed to and from a list the length of the list is also provided as a Nat

Methods

toList :: input -> [output]

> toList (Only 1)
  [1]

> toList (1,2,3)
  [1,2,3]

> toList ZeroTuple
  []

fromList' :: [output] -> input

a partial fromList with bad error messages

type ILL = IndexedListLiterals

An alias for IndexedListLiterals

module Data.Tuple.Only

data ZeroTuple a

Intuitively the zero tuple is () or Void but this breaks the Functional Dependency "input -> output length" stopping reliable inference, so this constructor is used to preserve type information

Constructors

ZeroTuple

len :: Proxy a

fromList :: forall input (length :: Nat) output. (KnownNat length, ILL input length output) => [output] -> Maybe input

> fromList [1,2,3] :: Maybe (Int, Int, Int)
  Just (1,2,3)

> fromList ["word","up"] :: Maybe (String, String, String)
  Nothing

> fromList ['z'] :: Maybe (Only Char)
  Just (Only 'z')

fromListP :: forall input (length :: Nat) output len. (KnownNat length, ILL input length output) => len length -> [output] -> Maybe input

the fromList variants take a list and convert it into a tuple it's sort of the inverse of toList

> fromListP (len @3) [1,2,3]
  Just (1,2,3)

> fromListP (len @3) ["word","up"]
  Nothing

> fromListP (len @1) ['z']
  Just (Only 'z') @